refactor(ivy): ngcc - move decorator analysis types into their own file
