+++ packages/compiler-cli/ngcc/test/rendering/dts_renderer_spec.ts
@@ -11,7 +11,8 @@ import {absoluteFrom, getFileSystem} from '../../../src/ngtsc/file_system';
 import {TestFile, runInEachFileSystem} from '../../../src/ngtsc/file_system/testing';
 import {loadTestFiles} from '../../../test/helpers';
 import {Import, ImportManager} from '../../../src/ngtsc/translator';
-import {CompiledClass, DecorationAnalyzer} from '../../src/analysis/decoration_analyzer';
+import {DecorationAnalyzer} from '../../src/analysis/decoration_analyzer';
+import {CompiledClass} from '../../src/analysis/types';
 import {NgccReferencesRegistry} from '../../src/analysis/ngcc_references_registry';
 import {ModuleWithProvidersAnalyzer, ModuleWithProvidersInfo} from '../../src/analysis/module_with_providers_analyzer';
 import {PrivateDeclarationsAnalyzer, ExportInfo} from '../../src/analysis/private_declarations_analyzer';
